requirements.txt
Flask==3.0.3
flask==2.3.3
python-dotenv==1.0.0
requests==2.32.0
flask-cors==4.0.0